actions-workflows: Initial shfmt action and workflow
Our current shfmt workflow steps are a little fragile, as we install a
go version then download the prebuilt shfmt binary. If there is a
mismatch between the installed go version and the go version expected by
the shfmt binary, we get cryptic error failures and the checks do not
work properly, which can frustrate developers trying to submit pull
requests to us.

To resolve this, unify our shfmt workflow and turn it into a proper
action that runs from a Docker file. As a result, we can install shfmt
from Alpine Linux's edge version, which is rolling release and has the
latest version most of the time, without needing to install go. Alpine
Linux was designed for containers so it is super fast to run and
install, meaning we sacrifice nothing to keep the check working.

fd is used as a simpler find replacement.

After this has been merged, we can go through every repository and
change the shfmt steps to just be:

uses: ClangBuiltLinux/actions-workflows/.github/workflows/shfmt.yml@main

Signed-off-by: Nathan Chancellor <nathan@kernel.org>