workflows: python_lint: Add support for linting with ruff
ruff is a fast Python linter written in Rust. Add support for linting
with ruff if a ruff.toml file is found within a repository. This allows
projects to have a set of warnings per project.

ruff can basically replace flake8 with a proper configuration, so do not
bother running flake8 if ruff is run.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>